Update existing score threshold comment
